# docs/openapi.yaml
components:
  schemas:
...
    CreateScheduleRequest:
      example:
        cron_expr: 0 2 * * *
        name: nightly-fleet
        tags:
          - prod
          - critical
        template:
          input: list:fleet.txt
          plugins:
            - modbus
            - s7
            - bacnet
        timezone: Europe/Madrid
      properties:
        audit_retention_days:
          description: v1.89+ per-schedule retention override. 0=inherit global, >0=keep N days.
          maximum: 3650
          minimum: 0
          type: integer
        cron_expr:
          description: 'Cadence: cron-based (5-field). Mutually exclusive with interval_seconds.'
          example: 0 2 * * *
          type: string
        interval_seconds:
          description: 'Cadence: interval-based. Mutually exclusive with cron_expr.'
          maximum: 604800
          minimum: 60
          type: integer
        name:
          example: nightly-fleet
          type: string
        tags:
          description: v2.4+ grouping labels.
          items:
            pattern: ^[a-z0-9_-]{1,32}$
            type: string
          maxItems: 10
          type: array
        template:
          $ref: '#/components/schemas/SubmitRequest'
        timezone:
          description: IANA zone for cron evaluation. Ignored for interval schedules.
          example: Europe/Madrid
          type: string
      required:
        - name
        - template
      type: object
    Envelope:
      properties:
        data: {}
        schema:
          example: api:v1
          type: string
      required:
        - schema
        - data
      type: object
...
    ScanSchedule:
      example:
        audit_retention_days: 30
        created_at: "2026-05-15T08:00:00Z"
        enabled: true
        id: abc123def4567890
        interval_seconds: 3600
        name: nightly-fleet
        next_fire_at: "2026-05-18T09:00:00Z"
        operator: alice@example.com
        tags:
          - critical
          - prod
        template:
          input: list:fleet.txt
          plugins:
            - modbus
            - s7
            - bacnet
        timezone: ""
        updated_at: "2026-05-15T08:00:00Z"
      properties:
        audit_retention_days:
          description: v1.89+ per-schedule retention override; 0 = inherit global.
          example: 30
          type: integer
        created_at:
          format: date-time
          type: string
        cron_expr:
          example: 0 2 * * *
          type: string
        enabled:
          example: true
          type: boolean
        id:
          example: abc123def4567890
          type: string
        interval_seconds:
          example: 3600
          type: integer
        last_fired_at:
          format: date-time
          type: string
        name:
          example: nightly-fleet
          type: string
        next_fire_at:
          description: v1.77+ computed; not persisted.
          format: date-time
          type: string
        operator:
          example: alice@example.com
          type: string
        tags:
          description: 'v2.4+ grouping labels (canonicalised: sorted + deduped).'
          example:
            - prod
            - critical
          items:
            type: string
          type: array
        template:
          $ref: '#/components/schemas/SubmitRequest'
        timezone:
          example: Europe/Madrid
          type: string
        updated_at:
          format: date-time
          type: string
      required:
        - id
        - name
        - template
        - enabled
        - created_at
        - updated_at
      type: object
...
    SubmitRequest:
      properties:
        default_port:
          type: integer
        input:
          example: list:fleet.txt
          type: string
        plugins:
          items:
            type: string
          type: array
      required:
        - input
      type: object
    UpdateScheduleRequest:
      properties:
        audit_retention_days:
          maximum: 3650
          minimum: 0
          type: integer
        cron_expr:
          type: string
        interval_seconds:
          maximum: 604800
          minimum: 60
          type: integer
        name:
          type: string
        tags:
          items:
            pattern: ^[a-z0-9_-]{1,32}$
            type: string
          maxItems: 10
          type: array
        template:
          $ref: '#/components/schemas/SubmitRequest'
        timezone:
          type: string
      required:
        - name
        - template
      type: object
info:
  description: Read-only HTTP API for ElSereno's dashboard. All endpoints return JSON with a `schema` discriminator so downstream tools can version against a concrete contract. Bearer token auth (ADR-014) gates /api/v1/* when the dashboard's auth mode is enabled.
  title: ElSereno HTTP API
  version: dev
openapi: 3.1.0
paths:
  /api/v1/findings:
    get:
      description: 'Query params: severity, protocol, min_score, created_after (RFC3339), limit (clamped 1..500, default 50). Returned order is created_at DESC; paginate by passing the oldest returned row''s created_at back as created_after.'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Envelope payload.
        "503":
          description: DB backend unavailable.
      summary: List findings (DB-backed, cursor-paginated).
      tags:
        - findings
...
  /api/v1/inputs/preview:
    get:
      description: 'Query params: kind=list:<path>|nmap:<path>|stdin (required), default_port=<int> (optional, host-only entries fall back here). Returns {count, targets[], truncated}; sample is capped at 200 entries. Provider kinds (shodan/censys/fofa/zoomeye/onyphe/internetdb) are NOT in scope here, they need credentials + rate-limit tuning that the dashboard process intentionally doesn''t carry; use the CLI scan verb for those. v1.36+.'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Envelope payload.
        "400":
          description: Missing kind, unsupported kind (provider kinds are rejected), or bad default_port.
        "404":
          description: 'list:/nmap: path doesn''t exist.'
        "500":
          description: Parse failure inside the underlying parser.
      summary: Preview targets parsed from an input source (no scan).
      tags:
        - inputs
...
  /api/v1/runs:
    get:
      description: 'Query params: status, started_after (RFC3339), limit (1..100, default 20).'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Envelope payload.
        "503":
          description: DB backend unavailable.
      summary: List recent scan runs with per-run finding counts.
      tags:
        - runs
  /api/v1/schedules:
    get:
      description: 'Returns every schedule sorted by name. `next_fire_at` is computed per response (v1.77+). v2.4+: optional `?tag=<value>` filters via array-overlap on the tags column.'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Envelope payload.
      summary: List scan schedules (v1.70+, v2.4 ?tag= filter).
      tags:
        - schedules
    post:
      description: Cadence is mutually-exclusive interval_seconds OR cron_expr (v1.73+). Optional timezone (IANA, v1.75+). Optional audit_retention_days (v1.89+) per-schedule retention override (0=inherit global).
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanSchedule'
          description: Schedule created.
        "400":
          description: Validation error (name/template/cadence/cron/tz/retention).
        "503":
          description: Schedule store unavailable.
      summary: Create a scan schedule.
      tags:
        - schedules
  /api/v1/schedules/{id}:
    delete:
      responses:
        "204":
          description: Deleted.
        "404":
          description: Schedule not found.
        "503":
          description: Schedule store unavailable.
      summary: Delete a schedule. Audit row persists with schedule_id NULL (v1.88+).
      tags:
        - schedules
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanSchedule'
          description: Schedule.
        "404":
          description: Schedule not found.
        "503":
          description: Schedule store unavailable.
      summary: Get one schedule.
      tags:
        - schedules
    put:
      description: 'Optional If-Match header (RFC3339 of stored updated_at; v1.78+) → 412 on mismatch. Optional X-Schedule-Force-Overwrite: true → audit force-overwrite event (v1.84+).'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScheduleRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanSchedule'
          description: Updated schedule.
        "400":
          description: Validation error.
        "404":
          description: Schedule not found.
        "412":
          description: Optimistic-locking precondition failed (If-Match mismatch).
        "503":
          description: Schedule store unavailable.
      summary: Update a schedule (v1.74+).
      tags:
        - schedules
  /api/v1/schedules/{id}/audit:
    get:
      description: 'Returns events newest-first. event_types: force_overwrite, delete, set_enabled_true, set_enabled_false. data is an array of ScheduleAuditEvent.'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: 'Envelope with data: ScheduleAuditEvent[].'
        "404":
          description: Schedule not found.
        "503":
          description: Audit store unavailable.
      summary: List audit events for a schedule (v1.84+).
      tags:
        - schedules
  /api/v1/schedules/{id}/clone:
    post:
      description: Body fields override the source. Default name is `<source.name> (copy)`. Clone always starts Enabled=true; LastFiredAt reset; operator = the cloner.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloneScheduleRequest'
        description: Optional override fields; empty body = full copy.
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanSchedule'
          description: Clone created.
        "400":
          description: Validation error.
        "404":
          description: Source schedule not found.
        "503":
          description: Schedule store unavailable.
      summary: Duplicate a schedule (v1.93+).
      tags:
        - schedules
  /api/v1/schedules/{id}/clones:
    get:
      description: 'Query param: depth (1..10, default 1; v2.23+). depth=1 returns direct clones only; depth=N walks the chain N hops. Result rows carry `clone_depth` (1..N) so the UI can render an ASCII tree. 404 when the source schedule itself is unknown.'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: 'Envelope with data: ScanSchedule[] (clone_depth populated).'
        "400":
          description: Malformed ?depth (must be 1..10).
        "404":
          description: Source schedule not found.
        "503":
          description: Schedule store unavailable.
      summary: List clones of a schedule (v2.10+, v2.23 depth).
      tags:
        - schedules
  /api/v1/schedules/{id}/disable:
    post:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Envelope payload.
        "404":
          description: Schedule not found.
        "503":
          description: Schedule store unavailable.
      summary: Disable a schedule. Writes audit (v1.88+).
      tags:
        - schedules
  /api/v1/schedules/{id}/enable:
    post:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Envelope payload.
        "404":
          description: Schedule not found.
        "503":
          description: Schedule store unavailable.
      summary: Enable a schedule. Writes audit (v1.88+).
      tags:
        - schedules
  /api/v1/schedules/{id}/runs:
    get:
      description: 'Query params: limit (1..1000, default 50), before (RFC3339 cursor). v2.0+: response is now {items: Job[], next_before?: rfc3339}. Paginate by reading next_before + passing it back as ?before=. Empty next_before ≡ last page. v2.45+: optional since=<rfc3339> + until=<rfc3339> for time-window filter; mutually exclusive with cursor (window supersedes when set). Inverted window (until<since) → 400.'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: '{items, next_before?} envelope.'
        "400":
          description: Malformed ?before / ?since / ?until; or inverted range.
        "404":
          description: Schedule not found.
        "503":
          description: Scan store unavailable.
      summary: List scheduler-fired jobs for a schedule (v1.92+, v2.0 cursor, v2.45 window).
      tags:
        - schedules
  /api/v1/schedules/{id}/stats:
    get:
      description: 'Query param: days (1..365, default 7). Returns total_runs / per-state counters / success_rate / avg_duration_seconds / avg_findings_per_run / total_findings over the window.'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Aggregate stats payload.
        "400":
          description: Malformed ?days query.
        "404":
          description: Schedule not found.
        "503":
          description: Scan store unavailable.
      summary: Aggregate run-stats for a schedule (v2.2+).
      tags:
        - schedules
  /api/v1/schedules/{id}/stats/timeseries:
    get:
      description: 'Query params: bucket=hour|day|week (default day), days=N (1..365, default 7). Returns {schedule_id, bucket, window_days, series: [{bucket_start, total_runs, completed, failed, cancelled, total_findings}]}. Memory store pre-fills empty buckets; PG store omits them.'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Envelope with timeseries payload.
        "400":
          description: Unknown ?bucket (must be hour|day|week).
        "404":
          description: Schedule not found.
        "503":
          description: Scan store unavailable.
      summary: Time-bucketed run-stats for a schedule (v2.11+).
      tags:
        - schedules
  /api/v1/schedules/audit:
    delete:
      description: Operator must opt into a cutoff explicitly. Returns {deleted_count, cutoff}.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Prune result.
        "400":
          description: Missing or malformed ?before.
        "503":
          description: Audit store unavailable.
      summary: Prune audit events older than ?before=<rfc3339> (v1.86+).
      tags:
        - schedules
  /api/v1/schedules/bulk/disable:
    post:
      description: No body. Returns count of state transitions (already-disabled schedules are no-ops). Used for planned-maintenance windows.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: '{affected, failed_audits, target_state}.'
        "503":
          description: Schedule store unavailable.
      summary: Bulk-disable every schedule (v1.95+).
      tags:
        - schedules
  /api/v1/schedules/bulk/enable:
    post:
      description: No body. Returns count of state transitions (already-enabled schedules are no-ops).
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: '{affected, failed_audits, target_state}.'
        "503":
          description: Schedule store unavailable.
      summary: Bulk-enable every schedule (v1.95+).
      tags:
        - schedules
  /api/v1/schedules/export:
    get:
      description: 'Query param: format=csv|ndjson|json|ics (default json). CSV is 10-column flat (id, name, cadence, enabled, operator, created_at, last_fired_at, audit_retention_days, input, plugins). NDJSON is round-trippable via per-line POST. v2.49+: ics emits RFC 5545 iCalendar, interval cadence → RRULE FREQ=HOURLY/DAILY best-fit; cron → DESCRIPTION-only; disabled → STATUS:CANCELLED. Content-Disposition: attachment with sensible filenames.'
      responses:
        "200":
          description: Body in the requested format.
        "400":
          description: Unsupported format.
        "503":
          description: Schedule store unavailable.
      summary: Export schedules for backup (v1.97+, v2.49 ics).
      tags:
        - schedules
  /api/v1/schedules/import:
    post:
      description: 'Body: NDJSON (one ScanSchedule per line; matches the v1.97 export format) OR JSON array. Content-Type drives the parser; unknown CT auto-detects via the first non-whitespace byte. Query param `on_conflict=skip|overwrite|rename` (default skip). IDs from the input are IGNORED, the server generates fresh IDs in every case. Returns aggregate counts + per-row items.'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Import result with per-row outcomes.
        "400":
          description: Malformed body or unsupported on_conflict value.
        "503":
          description: Schedule store unavailable.
      summary: Import schedules (v1.99+).
      tags:
        - schedules
...
  /api/v1/schedules/tags:
    get:
      description: Returns [{tag, count}] sorted by count DESC, tag ASC. Empty fleet → empty array.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: TagCount array.
        "503":
          description: Schedule store unavailable.
      summary: Aggregate tag-counts across the fleet (v2.5+).
      tags:
        - schedules
...
  /api/v1/stream:
    get:
      description: 'text/event-stream with `event:`, `id:`, `data:` lines. Event kinds: `finding`, `run_start`, `run_end`, `audit`. Payloads are JSON objects; schema per-kind is documented in internal/web/stream/findings_bridge.go.'
      responses:
        "200":
          description: SSE-framed event stream.
        "503":
          description: Live feed unavailable (broadcaster not wired).
      summary: Server-Sent Events feed (findings / runs / audit).
      tags:
        - stream
  /api/v1/triage:
    get:
      description: Returns an array of {severity, count} objects, ordered critical → info. Empty severities are omitted.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Envelope payload.
        "503":
          description: DB backend unavailable.
      summary: Per-severity finding counts for the triage panel.
      tags:
        - triage
...
servers:
  - url: http://127.0.0.1:8787